requirements.txt
fastapi>=0.109
uvicorn[standard]>=0.27
gunicorn>=21.2
pydantic>=2.5
pydantic-settings>=2.0
numpy>=1.26
requests>=2.31
shapely>=2.0
pyshp>=2.3
rich>=13.0
redis>=5.0
PyJWT>=2.8
cryptography>=41.0
supabase>=2.0
